+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 1.3.0 (2022-19-01)
+* Generic exception - add original exception class name to message [Issue 1] https://github.com/grasshopper7/rest-assured-extent-report-plugin/issues/1
+* Change newline pattern in exception parsing [Issue 2] https://github.com/grasshopper7/rest-assured-extent-report-plugin/issues/2
+
 ## 1.2.0 (2021-12-31)
 
 * Fixed functionality bugs